refactor: move create route under subdirectory
