--- ayusin-be/src/routes/projects/checks/index.ts
@@ -1,2 +1,0 @@
-export { addCheckHandler } from "./handlers";
-export { addCheckRoute } from "./routes";
